test(pr): split github client fallback tests for LOC limit
